requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
cloudinary==1.40.0
coreapi==2.3.3
coreschema==0.0.4
dj-database-url==2.1.0
Django==5.0.6
django-cors-headers==4.3.1
djangorestframework==3.15.1
drf-yasg==1.21.7
gunicorn==22.0.0
idna==3.7
inflection==0.5.1
install==1.3.5
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mysqlclient==2.2.4
openapi-codec==1.3.2
packaging==24.0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
requests==2.32.2
simplejson==3.19.2
six==1.16.0
sqlparse==0.5.0
typing_extensions==4.12.0
uritemplate==4.1.1
urllib3==2.2.1
psycopg2==2.9.9
psycopg2-binary==2.9.9
